Suggested rewrite for closestEvent.
The function uses foldr to compute the minimum of a list of events using a
non-standard ordering (namely the dist field of the Event type).

There is a much more direct way to express this:

> minimumBy :: (a -> a -> Ordering) -> [a] -> a

The function takes a comparison function as an argument. The default
comparison function is called "compare". We don't want to compare
events directly, but their dist fields. We can use another helpful function
from Data.Function:

> on :: (b -> b -> c) -> (a -> b) -> (a -> a -> c)

This function takes a binary operator on b's and turns it into a binary
operator on a's, given that we can go from a to b. In our setting, we
can use "on" to take "compare" on FP to a comparison on Events, because
we can extract an FP from events using "dist". All together, we get

> minimumBy (on compare dist)

Now, we can write "on" infix, and arrive at the version I've committed.