Adding a basic particle class which implements a new particle class.
The Particle class captures the basic interface of a particle
object. The environment and events are members of the type family. The
only method is step, which takes the environment, a particle and
results in an Event and new Particle (Maybe).

The Environment for BasicParticle is just an end of the timestep.