Fixing the space leak.
The main culprit was the "union" operation on finite maps, which isn't strict.
Strictness annotations on the finite map of type PhysicsTally don't help,
because they only force evaluation to WHNF of the complete map, which isn't
doing much for a complex data structure such as finite maps. I've preliminarily
"fixed" this by calling deepseq on the map during the iteration, which completely
evaluates the data structure at every step. It is worth investigating, though,
whether another data structure might be more adequate for PhysicsTally.

With this fix, we can make runManyP and runManyP6 run in constant space and
linear time (if -O is enabled). They key is to use a strict left fold (foldl')
for the iteration. Furthermore, in runManyP6, a pair is used in the history
function to keep two pieces of state during the iteration. This pair has to
be strict (that's why I inserted bang patterns) -- otherwise we get another
space leak.